--- a/velox/expression/VectorUdfTypeSystem.h
+++ b/velox/expression/VectorUdfTypeSystem.h
@@ -150,10 +150,13 @@ template <typename T, typename = void>
 struct VectorReader {
   using exec_in_t = typename VectorExec::template resolver<T>::in_type;
 
-  VectorReader(const DecodeResult<exec_in_t>& decoded) : decoded_{decoded} {}
+  explicit VectorReader(const DecodedVector* decoded) : decoded_(*decoded) {}
 
-  const exec_in_t& operator[](size_t offset) const {
-    return decoded_[offset];
+  explicit VectorReader(const VectorReader<T>&) = delete;
+  VectorReader<T>& operator=(const VectorReader<T>&) = delete;
+
+  exec_in_t operator[](size_t offset) const {
+    return decoded_.template valueAt<exec_in_t>(offset);
   }
 
   bool isSet(size_t offset) const {
@@ -166,14 +169,14 @@ struct VectorReader {
 
   bool doLoad(size_t offset, exec_in_t& v) const {
     if (isSet(offset)) {
-      v = decoded_[offset];
+      v = decoded_.template valueAt<exec_in_t>(offset);
       return true;
     } else {
       return false;
     }
   }
 
-  const DecodeResult<exec_in_t> decoded_;
+  const DecodedVector& decoded_;
 };
 
 template <typename K, typename V>
@@ -261,17 +264,16 @@ struct VectorWriter<Map<K, V>> {
 
 namespace detail {
 
-template <typename TOut, typename TIn>
-const TOut& getDecoded(const DecodeResult<TIn>& decoded) {
+template <typename TOut>
+const TOut& getDecoded(const DecodedVector& decoded) {
   auto base = decoded.base();
   return *base->template as<TOut>();
 }
 
-template <typename T>
-DecodeResult<T> decode(DecodedVector& decoder, const BaseVector& vector) {
+inline DecodedVector* decode(DecodedVector& decoder, const BaseVector& vector) {
   SelectivityVector rows(vector.size());
   decoder.decode(vector, rows);
-  return decoder.as<T>();
+  return &decoder;
 }
 } // namespace detail
 
@@ -279,22 +281,20 @@ template <typename K, typename V>
 struct VectorReader<Map<K, V>> {
   using in_vector_t = typename TypeToFlatVector<Map<K, V>>::type;
   using exec_in_t = typename VectorExec::template resolver<Map<K, V>>::in_type;
-  using exec_in_key_t = typename VectorExec::template resolver<K>::in_type;
-  using exec_in_val_t = typename VectorExec::template resolver<V>::in_type;
 
-  explicit VectorReader(const DecodeResult<exec_in_t>& decoded)
-      : decoded_{decoded},
+  explicit VectorReader(const DecodedVector* decoded)
+      : decoded_{*decoded},
         vector_(detail::getDecoded<in_vector_t>(decoded_)),
         offsets_(vector_.rawOffsets()),
         lengths_(vector_.rawSizes()),
-        keyReader_{
-            detail::decode<exec_in_key_t>(decodedKeys_, *vector_.mapKeys())},
-        valReader_{
-            detail::decode<exec_in_val_t>(decodedVals_, *vector_.mapValues())} {
-  }
+        keyReader_{detail::decode(decodedKeys_, *vector_.mapKeys())},
+        valReader_{detail::decode(decodedVals_, *vector_.mapValues())} {}
+
+  explicit VectorReader(const VectorReader<Map<K, V>>&) = delete;
+  VectorReader<Map<K, V>>& operator=(const VectorReader<Map<K, V>>&) = delete;
 
   exec_in_t operator[](size_t offset) const {
-    auto index = decoded_.decodedIndex(offset);
+    auto index = decoded_.index(offset);
     return MapView{&keyReader_, &valReader_, offsets_[index], lengths_[index]};
   }
 
@@ -302,7 +302,7 @@ struct VectorReader<Map<K, V>> {
     return !decoded_.isNullAt(offset);
   }
 
-  DecodeResult<exec_in_t> decoded_;
+  const DecodedVector& decoded_;
   const MapVector& vector_;
   DecodedVector decodedKeys_;
   DecodedVector decodedVals_;
@@ -320,13 +320,16 @@ struct VectorReader<Array<V>> {
   using exec_in_t = typename VectorExec::template resolver<Array<V>>::in_type;
   using exec_in_child_t = typename VectorExec::template resolver<V>::in_type;
 
-  explicit VectorReader(const DecodeResult<exec_in_t>& decoded)
-      : decoded_(decoded),
+  explicit VectorReader(const DecodedVector* decoded)
+      : decoded_(*decoded),
         vector_(detail::getDecoded<in_vector_t>(decoded_)),
         offsets_{vector_.rawOffsets()},
         lengths_{vector_.rawSizes()},
-        childReader_{
-            detail::decode<exec_in_child_t>(decoder_, *vector_.elements())} {}
+        childReader_{detail::decode(arrayValuesDecoder_, *vector_.elements())} {
+  }
+
+  explicit VectorReader(const VectorReader<Array<V>>&) = delete;
+  VectorReader<Array<V>>& operator=(const VectorReader<Array<V>>&) = delete;
 
   // TODO: Once other types are converted to view types, the doLoad protocol
   // will be removed.
@@ -343,12 +346,12 @@ struct VectorReader<Array<V>> {
   }
 
   exec_in_t operator[](size_t offset) const {
-    auto index = decoder_.index(offset);
+    auto index = arrayValuesDecoder_.index(offset);
     return ArrayView{&childReader_, offsets_[index], lengths_[index]};
   }
 
-  DecodedVector decoder_;
-  DecodeResult<exec_in_t> decoded_;
+  DecodedVector arrayValuesDecoder_;
+  const DecodedVector& decoded_;
   const ArrayVector& vector_;
   const vector_size_t* offsets_;
   const vector_size_t* lengths_;
@@ -435,11 +438,11 @@ struct VectorReader<Row<T...>> {
   using in_vector_t = typename TypeToFlatVector<Row<T...>>::type;
   using exec_in_t = typename VectorExec::resolver<Row<T...>>::in_type;
 
-  explicit VectorReader(const DecodeResult<exec_in_t>& decoded)
-      : decoded_(decoded),
+  explicit VectorReader(const DecodedVector* decoded)
+      : decoded_(*decoded),
         vector_(detail::getDecoded<in_vector_t>(decoded_)),
-        decoders_{vector_.childrenSize()},
-        childReader_{prepareChildReaders(
+        childrenDecoders_{vector_.childrenSize()},
+        childReaders_{prepareChildReaders(
             vector_,
             std::make_index_sequence<sizeof...(T)>{})} {}
 
@@ -464,19 +467,11 @@ struct VectorReader<Row<T...>> {
 
  private:
   template <size_t... I>
-  std::tuple<VectorReader<T>...> prepareChildReaders(
+  std::tuple<std::unique_ptr<VectorReader<T>>...> prepareChildReaders(
       const in_vector_t& vector,
       std::index_sequence<I...>) {
-    return {childReader<T, I>()...};
-  }
-
-  template <typename CHILD_T, size_t I>
-  VectorReader<CHILD_T> childReader() {
-    auto& decoder = decoders_[I];
-    return VectorReader<CHILD_T>(
-        detail::decode<
-            typename VectorExec::template resolver<CHILD_T>::in_type>(
-            decoder, *vector_.childAt(I)));
+    return {std::make_unique<VectorReader<T>>(
+        detail::decode(childrenDecoders_[I], *vector_.childAt(I)))...};
   }
 
   template <size_t N, typename Type, typename... Types>
@@ -485,7 +480,7 @@ struct VectorReader<Row<T...>> {
 
     exec_child_in_t childval{};
 
-    if (std::get<N>(childReader_).doLoad(offset, childval)) {
+    if (std::get<N>(childReaders_)->doLoad(offset, childval)) {
       results.template at<N>() = childval;
     }
 
@@ -494,10 +489,10 @@ struct VectorReader<Row<T...>> {
     }
   }
 
-  DecodeResult<exec_in_t> decoded_;
+  const DecodedVector& decoded_;
   const in_vector_t& vector_;
-  std::vector<DecodedVector> decoders_;
-  std::tuple<VectorReader<T>...> childReader_;
+  std::vector<DecodedVector> childrenDecoders_;
+  std::tuple<std::shared_ptr<VectorReader<T>>...> childReaders_;
   mutable exec_in_t returnval_;
 };