**Only last commit belongs to this PR**

DecodedResults is used in Vector Readers to access the values and the nullability of 
the underlying vector. It has some efficiencies, such as forcing indicies to materialize and 
not checking identity and constant mapping. Solving those made DecodedResults just a repetition 
of DecodedVector. 
in this diff i remove DecodedResults and use the underlying DecodedVector instead.

## Performance:
Its interesting that now, nestedMapSumSimpleFunction and nestedMapSumVectorFunctionMapView 
do not have similar performance, i expect the expanded vectorFunctionAdapter for the simple function 
to look like nestedMapSumVectorFunctionMapView.
after the diff:
```
============================================================================
mapSumVectorFunction                                         2.91ms   343.24
mapSumSimpleFunction                              89.23%     3.27ms   306.28
nestedMapSumVectorFunction                                   6.36ms   157.33
nestedMapSumSimpleFunction                        61.92%    10.27ms    97.42
nestedMapSumVectorFunctionMapView                 99.47%     6.39ms   156.50
============================================================================
```
before the diff:
```
============================================================================
mapSumVectorFunction                                         2.97ms   337.17
mapSumSimpleFunction                              40.09%     7.40ms   135.17
nestedMapSumVectorFunction                                   6.19ms   161.47
nestedMapSumSimpleFunction                        26.78%    23.12ms    43.25
nestedMapSumVectorFunctionMapView                 29.53%    20.97ms    47.68
============================================================================
```

one more point is the min.max: 
now simpleMinIntegerIterator is 17% slower than vectorMinInteger it used to be
30%slower.
```
vectorMinInteger                                  86.31%   557.59us    1.79K
vectorMinIntegerNoFastPath                        60.64%   793.60us    1.26K
simpleMinIntegerIterator                          69.31%   694.34us    1.44K
```