@laithsakka 2021-11-15
address comments related to commit 2